# docs/swagger.yaml
definitions:
  dto.AccountDeletionRes:
    properties:
      deleteionDate:
        example: 2023-12-11 03:21:18 +0000 UTC
        type: string
      keyID:
        example: f50a9229-e7c7-45ba-b06c-8036b894424e
        type: string
    type: object
  dto.AccountListRes:
    properties:
      accounts:
        items:
          $ref: '#/definitions/dto.AccountRes'
        type: array
      marker:
        example: AE0AAAACAHMAAAAJYWNjb3VudElkAHMAAAAMOTg1MDk2Mzk3ODIxAHMAAAAEdGtJZABzAAAAJDQ0YTAzNWU2LTY1OTEtNDgwMC04YjcwLWM3MzNiNTI2MzljMw
        type: string
    type: object
  dto.AccountRes:
    properties:
      address:
        example: 0x216690cD286d8a9c8D39d9714263bB6AB97046F3
        type: string
      keyID:
        example: f50a9229-e7c7-45ba-b06c-8036b894424e
        type: string
    type: object
  dto.PkReq:
    properties:
      pk:
        example: 637081577126ff5c2d327f992bd66548e00262578fd558d7fe272ef21b8bf825
        type: string
    required:
    - pk
    type: object
  dto.SingedTxnRes:
    properties:
      signedTxn:
        example: 0xf86a5685ba43b740008252089439e243a7f209932df41e1fc0a1ada51b3a04b46d0180860b280f5b1d3aa00d2ea43cfd9b91151348d037a5a80293f543e1700a7019853f28063f6442c826a052a29797169740b1bc48962e197299061c8aa3314951a9c71418d19036604645
        type: string
    type: object
  dto.TxnReq:
    properties:
      keyID:
        example: f50a9229-e7c7-45ba-b06c-8036b894424e
        type: string
      serializedTxn:
        example: 0xea5685ba43b740008252089439e243a7f209932df41e1fc0a1ada51b3a04b46d018086059407ad8e8b8080
        type: string
    required:
    - keyID
    - serializedTxn
    type: object
info:
  contact: {}
paths:
  /api/accounts:
    get:
      parameters:
      - example: 100
        in: query
        maximum: 1000
        name: limit
        type: integer
      - in: query
        name: marker
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AccountListRes'
      summary: Get accounst list
      tags:
      - Kms
  /api/accounts/{keyID}:
    delete:
      parameters:
      - description: kms key-id
        in: path
        name: keyID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AccountDeletionRes'
      summary: delete account of target key id
      tags:
      - Kms
    get:
      parameters:
      - description: kms key-id
        in: path
        name: keyID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AccountRes'
      summary: Get account of target key id
      tags:
      - Kms
  /api/create/account:
    post:
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.AccountRes'
      summary: Create new account
      tags:
      - Kms
  /api/error:
    get:
      responses:
        "200":
          description: OK
      tags:
      - Health
  /api/health:
    get:
      responses:
        "200":
          description: OK
      tags:
      - Health
  /api/import/account:
    post:
      parameters:
      - description: subject
        in: body
        name: subject
        required: true
        schema:
          $ref: '#/definitions/dto.PkReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.AccountRes'
      summary: Import account to kms
      tags:
      - Kms
  /api/sign/txn:
    post:
      parameters:
      - description: subject
        in: body
        name: subject
        required: true
        schema:
          $ref: '#/definitions/dto.TxnReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.SingedTxnRes'
      summary: Sign serialized transaction.
      tags:
      - Transaction
swagger: "2.0"